Move default provider logic to new file
